Move Psycopg2 into its own submodule
